+++ src/libft/ft_uatoi_base.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_atoi_base.c                                     :+:      :+:    :+:   */
+/*   ft_uatoi_base.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: scoron <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/11 18:52:00 by scoron            #+#    #+#             */
-/*   Updated: 2019/01/31 14:04:02 by scoron           ###   ########.fr       */
+/*   Updated: 2020/05/09 20:52:48 by scoron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	ft_check_error_str(char *str, char *base)
+static int			ft_check_error_str(char *str, char *base)
 {
 	int i;
 	int j;
@@ -39,7 +39,7 @@ static int	ft_check_error_str(char *str, char *base)
 	return (1);
 }
 
-static int	ft_check_error_b(char *base)
+static int			ft_check_error_b(char *base)
 {
 	int i;
 	int j;
@@ -64,7 +64,7 @@ static int	ft_check_error_b(char *base)
 	return (0);
 }
 
-static int	ft_find(char *str, char *base, int i, int j)
+static int			ft_find(char *str, char *base, int i, int j)
 {
 	int r;
 
@@ -92,11 +92,11 @@ static int	ft_find(char *str, char *base, int i, int j)
 	}
 }
 
-uintmax_t	ft_uatoi_base(char *str, char *base)
+unsigned long long	ft_uatoi_base(char *str, char *base)
 {
 	int					bs;
 	int					i;
-	uintmax_t			nb;
+	unsigned long long	nb;
 
 	bs = 0;
 	nb = 0;
